rewrite handleUpload
This way we avoid the right-drift that we introduced in the first
version, as the error handling and propagation is done in the Maybe
monad.

The downside to this is that we lose some error information: We cannot
show anymore if the authorization failed or something else went wrong,
because in all cases we just bail out with Nothing.